@@ -8,10 +8,10 @@ class EnsureEmailFilledTest extends TestCase
 {
     public function testHandle()
     {
-        $noEmailUser = factory(User::class)->make(['email' => '']);
+        $noEmailUser = User::factory()->make(['email' => '']);
         $this->actingAs($noEmailUser)->get('/user')->assertRedirect('/auth/bind');
 
-        $normalUser = factory(User::class)->make();
+        $normalUser = User::factory()->make();
         $this->actingAs($normalUser)->get('/auth/bind')->assertRedirect('/user');
     }
 }
