@@ -10,7 +10,7 @@ class FireUserAuthenticatedTest extends TestCase
     public function testHandle()
     {
         Event::fake();
-        $user = factory(User::class)->make();
+        $user = User::factory()->make();
         $this->actingAs($user)->get('/user');
         Event::assertDispatched(\App\Events\UserAuthenticated::class, function ($event) use ($user) {
             $this->assertEquals($user->uid, $event->user->uid);
