@@ -15,7 +15,7 @@ class SideMenuComposerTest extends TestCase
     {
         Event::fake();
 
-        $admin = factory(User::class)->states('admin')->create();
+        $admin = User::factory()->admin()->create();
         $this->actingAs($admin)->get('/user');
         Event::assertDispatched(Events\ConfigureUserMenu::class);
         Event::assertDispatched(Events\ConfigureExploreMenu::class);
@@ -28,7 +28,7 @@ class SideMenuComposerTest extends TestCase
     {
         $filter = Fakes\Filter::fake();
 
-        $user = factory(User::class)->create();
+        $user = User::factory()->create();
         $this->actingAs($user)->get('/user');
         $filter->assertHaveBeenApplied('side_menu', function ($menu, $type) {
             $this->assertCount(count(config('menu.user')), $menu);
@@ -46,7 +46,7 @@ class SideMenuComposerTest extends TestCase
 
     public function testTransform()
     {
-        $user = factory(User::class)->create();
+        $user = User::factory()->create();
         $this->actingAs($user);
 
         $crawler = new Crawler($this->get('/user/oauth/manage')->getContent());
@@ -73,7 +73,7 @@ class SideMenuComposerTest extends TestCase
                 );
         });
 
-        $admin = factory(User::class)->states('admin')->create();
+        $admin = User::factory()->admin()->create();
         $this->actingAs($admin)
             ->get('/admin')
             ->assertDontSee(trans('general.plugin-configs'));
