@@ -15,7 +15,7 @@ class HookTest extends TestCase
             'icon' => 'fa-book',
             'new-tab' => true,
         ]);
-        $this->actingAs(factory(User::class)->create())
+        $this->actingAs(User::factory()->create())
             ->get('/user')
             ->assertSee('Link A')
             ->assertSee('/to/a')
@@ -28,7 +28,7 @@ class HookTest extends TestCase
             'link' => '/to/b',
             'icon' => 'fa-book',
         ]);
-        $this->actingAs(factory(User::class)->create())
+        $this->actingAs(User::factory()->create())
             ->get('/user')
             ->assertSee('Link B')
             ->assertSee('/to/b');
@@ -73,14 +73,14 @@ class HookTest extends TestCase
     public function testAddUserBadge()
     {
         Hook::addUserBadge('hi', 'green');
-        $this->actingAs(factory(User::class)->create())
+        $this->actingAs(User::factory()->create())
             ->get('/user')
             ->assertSee('<span class="badge bg-green mb-1 mr-2">hi</span>', false);
     }
 
     public function testSendNotification()
     {
-        $user = factory(User::class)->create();
+        $user = User::factory()->create();
         Hook::sendNotification([$user], 'Ibara Mayaka');
         $user->refresh();
         $this->assertCount(1, $user->unreadNotifications);
