@@ -8,7 +8,7 @@ class HandlerTest extends TestCase
 {
     public function testRenderAjaxException()
     {
-        $json = $this->get('/abc', ['Accept' => 'application/json'])->decodeResponseJson();
+        $json = $this->get('/abc', ['Accept' => 'application/json'])->json();
         $this->assertIsString($json['message']);
         $this->assertTrue($json['exception']);
         $this->assertTrue(collect($json['trace'])->every(
